Se implementaron mejoras en cuanto a obtener informacion de la API. Mejoras de diseño.

+++ src/pages/Dashboard.jsx
@@ -1,37 +1,50 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import "../css/dashboard.css";
+import { AuthContext } from "../context/AuthContext";
 
 const Dashboard = () => {
   const [posts, setPosts] = useState([]); // Estado para almacenar los destinos
+  const { user } = useContext(AuthContext); // Usuario autenticado desde el contexto
   const navigate = useNavigate();
 
   useEffect(() => {
-    const fetchPosts = async () => {
+    const fetchUserPosts = async () => {
       try {
-        const response = await fetch(
-          "https://67391e4ea3a36b5a62edfb6e.mockapi.io/users/1/posts"
-        );
-        const data = await response.json();
-        setPosts(data);
+        if (user && user.id) {
+          const response = await fetch(
+            `https://67391e4ea3a36b5a62edfb6e.mockapi.io/users/${user.id}/posts`
+          );
+          const data = await response.json();
+          setPosts(data); // Asignar solo los posts del usuario autenticado
+        }
       } catch (error) {
         console.error("Error al cargar los destinos:", error);
       }
     };
 
-    fetchPosts();
-  }, []);
+    fetchUserPosts();
+  }, [user]);
 
   const handleDelete = async (id) => {
     try {
-      await fetch(`https://67391e4ea3a36b5a62edfb6e.mockapi.io/users/1/posts/${id}`, {
-        method: "DELETE",
-      });
-      setPosts(posts.filter((post) => post.id !== id)); // Actualiza la lista
+      if (user && user.id) {
+        await fetch(
+          `https://67391e4ea3a36b5a62edfb6e.mockapi.io/users/${user.id}/posts/${id}`,
+          {
+            method: "DELETE",
+          }
+        );
+        setPosts(posts.filter((post) => post.id !== id)); // Actualiza la lista
+      }
     } catch (error) {
       console.error("Error al eliminar el destino:", error);
     }
   };
+
+  if (!user) {
+    return <p>No tienes permisos para acceder a esta página.</p>;
+  }
 
   return (
     <div className="dashboard-container">
@@ -53,7 +66,10 @@ const Dashboard = () => {
               <button onClick={() => navigate(`/edit-destination/${post.id}`)}>
                 Editar
               </button>
-              <button onClick={() => handleDelete(post.id)} className="delete-btn">
+              <button
+                onClick={() => handleDelete(post.id)}
+                className="delete-btn"
+              >
                 Eliminar
               </button>
             </div>
